feat(cli)!: remove deprecated sync-deps command
BREAKING CHANGE: The `airis sync-deps` command has been removed.

- Move version resolution functions to `src/version_resolver.rs`
- Remove `src/commands/sync_deps.rs` entirely
- Update imports in diff, doctor, generate, and templates modules
- Update CHANGELOG.md with migration guide
- Update CLAUDE.md documentation

Users should use `airis init` or `airis generate files` instead,
which resolve catalog versions automatically.
