+++ src/commands/doctor.rs
@@ -12,7 +12,7 @@ use std::fs;
 use std::path::Path;
 
 use crate::commands::manifest_cmd::WorkspaceTruth;
-use crate::commands::sync_deps::resolve_version;
+use crate::version_resolver::resolve_version;
 use crate::manifest::{CatalogEntry, Manifest, MANIFEST_FILE};
 use crate::ownership::{get_ownership, Ownership};
 use crate::templates::TemplateEngine;
