+++ src/commands/mod.rs
@@ -4,7 +4,6 @@ pub mod bundle;
 pub mod clean;
 pub mod deps;
 pub mod diff;
-pub mod upgrade;
 pub mod discover;
 pub mod doctor;
 pub mod docs;
@@ -20,6 +19,6 @@ pub mod new_cmd;
 pub mod policy;
 pub mod run;
 pub mod shim;
-pub mod sync_deps;
+pub mod upgrade;
 pub mod validate_cmd;
 pub mod verify;
